Update install documentation and dependencies
Problem:

After cloning the repo and following the instructions in the README,
the application was throwing a bunch of errors when I tried to run it.

It turns out that several steps were missing from the README,
several files were missing from the repo,
and the documentation around setting up the github application
was ambiguous.

Solution:

- Add missing setup steps to the README, including `npm install`
- Clarify the configuration for the Github developer application
- Add empty files that are expected by `npm build`
- Add missing CSS variable